test: move mappers and DTOs to application layer
